remove shiny, get rstudio server running
rsession.conf has to be readable by the rstudio-server process. Current
solution is to make it world-readable, which is obviously dumb, but who
cares since it's just the default empty config.

Also changed the default username/pw from shiny/shiny to litdata/litdata

And stripped down the list of packages to install, for now.